+++ src/automyte/utils/random.py
@@ -2,5 +2,6 @@ import hashlib
 import uuid
 
 
-def random_hash():
-    return hashlib.md5(uuid.uuid4().bytes).hexdigest()
+def random_hash(of: str | None = None):
+    hash_from = of.encode() if of else uuid.uuid4().bytes
+    return hashlib.md5(hash_from).hexdigest()